backend/requirements.txt
# Web Framework
Django>=5.0,<5.2
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Database & Utils
psycopg2-binary>=2.9.9
sqlparse>=0.5.0
asgiref>=3.8.1
python-dotenv>=1.0.1
requests>=2.25.1

# Telegram & Crypto
Telethon>=1.42.0
pycryptodome>=3.20.0

# Production Server
gunicorn>=22.0.0

# Additional utilities
requests>=2.31.0
netifaces>=0.11.0
